+++ src/ideabox/policy/content/state_progress.py
@@ -1,9 +1,10 @@
 # -*- coding: utf-8 -*-
 from ideabox.policy import _
+from plone import api
 from plone.app.textfield import RichText
-from plone.dexterity.browser import view
 from plone.dexterity.content import Container
 from plone.supermodel import model
+from Products.Five import BrowserView
 from zope import schema
 from zope.interface import implementer
 
@@ -21,5 +22,8 @@ class StateProgress(Container):
     pass
 
 
-class StateProgressView(view.DefaultView):
-    pass
+class StateProgressView(BrowserView):
+    def get_state_progress(self):
+        return api.content.find(
+            contect=self, portal_type="state_progress", sort_on="state_date"
+        )